GH-21 # Cleanup
